refactor: change file structure
1. Remove out-of-date code archive
2. Update code file by moving code in "/" to pkg folder
