+++ src/tg_reviewer_bot/submit.py
@@ -1,11 +1,5 @@
-from telegram import (
-    InlineKeyboardButton,
-    InlineKeyboardMarkup,
-    InputMediaDocument,
-    InputMediaPhoto,
-    InputMediaVideo,
-    Update,
-)
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Update
+from telegram.constants import MessageOriginType
 from telegram.ext import (
     CallbackQueryHandler,
     CommandHandler,
@@ -14,10 +8,11 @@ from telegram.ext import (
     MessageHandler,
     filters,
 )
+from telegram.helpers import escape_markdown
 
-from const import TG_REVIEWER_GROUP
-from review import reply_review_message
-from utils import send_submission
+from db_op import Banned_user, Submitter
+from review_utils import reply_review_message
+from utils import TG_BANNED_NOTIFY, TG_REVIEWER_GROUP, send_submission
 
 # set const as the state of one user
 COLLECTING = range(1)
@@ -41,7 +36,13 @@ async def confirm_submission(
         await query.edit_message_text(text="投稿已取消")
     elif query.data.startswith(("anonymous", "realname")):
         if query.data.startswith("realname"):
-            submission["text"] += f"\n\nby {user.full_name}"
+            sign_string = f"_via_ [{escape_markdown(user.full_name,version=2,)}](tg://user?id={user.id})"
+            # if the last line is a forward message, put in the same line
+            if submission["text"].split("\n")[-1].startswith("_from_"):
+                submission["text"] += " " + sign_string
+            else:
+                submission["text"] += "\n\n" + sign_string
+
         submission_messages = await send_submission(
             context=context,
             chat_id=TG_REVIEWER_GROUP,
@@ -49,7 +50,7 @@ async def confirm_submission(
             media_type_list=submission["media_type_list"],
             documents_id_list=submission["document_id_list"],
             document_type_list=submission["document_type_list"],
-            text=submission["text"],
+            text=submission["text"].strip(),
         )
         submission_meta = {
             "submitter": [
@@ -59,6 +60,7 @@ async def confirm_submission(
                 submission["first_message_id"],
             ],
             "reviewer": {},
+            "text": submission['text'],
             "media_id_list": submission["media_id_list"],
             "media_type_list": submission["media_type_list"],
             "documents_id_list": submission["document_id_list"],
@@ -69,6 +71,7 @@ async def confirm_submission(
         await query.edit_message_text(text="投稿成功")
 
     del message_groups[user.id]
+    Submitter.count_increase(user.id, "submission_count")
     return ConversationHandler.END
 
 
@@ -86,22 +89,57 @@ async def collect_data(update: Update, context: ContextTypes.DEFAULT_TYPE):
     submission["last_confirm_button"] = None
 
     # add new message to this user's message_groups dict
-    if message.text:
-        submission["text"] += message.text + "\n"
-    if message.caption:
-        submission["text"] += message.caption + "\n"
+    if message.text_markdown_v2_urled:
+        submission["text"] += "\n\n" + message.text_markdown_v2_urled
+    if message.caption_markdown_v2_urled:
+        submission["text"] += "\n\n" + message.caption_markdown_v2_urled
     if message.photo:
         submission["media_id_list"].append(message.photo[-1].file_id)
         submission["media_type_list"].append("photo")
     if message.video:
         submission["media_id_list"].append(message.video.file_id)
         submission["media_type_list"].append("video")
-    if message.document:
+    if message.sticker:
+        submission["media_id_list"].append(message.sticker.file_id)
+        submission["media_type_list"].append("sticker")
+    if message.animation:   # GIF
+        submission["media_id_list"].append(message.animation.file_id)
+        submission["media_type_list"].append("animation")
+    # elif because gif is also a document but can not be sent as a group
+    elif message.document:
         submission["document_id_list"].append(message.document.file_id)
         submission["document_type_list"].append("document")
     if submission["first_message_id"] is None:
         submission["first_message_id"] = message.message_id
-
+    if message.forward_origin is not None:
+        forward_string = "\n\n_from_ "
+        match message.forward_origin.type:
+            case MessageOriginType.USER:
+                forward_name = message.forward_origin.sender_user.full_name
+                forward_url = (
+                    f"tg://user?id={message.forward_origin.sender_user.id}"
+                )
+                forward_string += f"[{escape_markdown(forward_name,version=2,)}]({forward_url})"
+            case MessageOriginType.CHAT:
+                forward_name = message.forward_origin.sender_chat.title
+                forward_url = message.forward_origin.sender_chat.link
+                forward_string += f"[{escape_markdown(forward_name,version=2,)}]({forward_url})"
+            case MessageOriginType.CHANNEL:
+                forward_name = message.forward_origin.chat.title
+                forward_url = f"{message.forward_origin.chat.link}/{message.forward_origin.message_id}"
+                forward_string += f"[{escape_markdown(forward_name,version=2,)}]({forward_url})"
+            case MessageOriginType.HIDDEN_USER:
+                forward_name = message.forward_origin.sender_user_name
+                forward_string += escape_markdown(
+                    forward_name,
+                    version=2,
+                )
+        if (
+            submission["last_origin"] == ""
+            or submission["last_origin"] != forward_name
+        ):
+            submission["text"] += forward_string
+        submission["last_origin"] = forward_name
     # show preview of all messages this user has sent
     submission["last_preview_messages"].extend(
         await send_submission(
@@ -135,7 +173,14 @@ async def collect_data(update: Update, context: ContextTypes.DEFAULT_TYPE):
 async def handle_new_submission(
     update: Update, context: ContextTypes.DEFAULT_TYPE
 ):
-    await update.message.reply_text("请开始你的投稿，你可以发送多条消息，包括文本和媒体。")
+    if Banned_user.is_banned(update.effective_user.id):
+        if TG_BANNED_NOTIFY:
+            await update.message.reply_text("你已被禁止投稿。")
+        return ConversationHandler.END
+
+    await update.message.reply_text(
+        "请开始你的投稿，你可以发送多条消息，包括文本和媒体。"
+    )
     # empty this user's message_groups dict
     message_groups[update.effective_user.id] = {
         "media_id_list": [],
@@ -148,12 +193,15 @@ async def handle_new_submission(
         "last_preview_messages": [],
         "last_confirm_button": None,
         "first_message_id": None,
+        "last_origin": "",
     }
     return COLLECTING
 
 
 async def err_input(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    await update.message.reply_text("错误的命令。\n如果想要重新投稿，请输入 /new")
+    await update.message.reply_text(
+        "错误的命令。\n如果想要重新投稿，请输入 /new"
+    )
     return COLLECTING
