+++ src/tg_reviewer_bot/utils.py
@@ -1,11 +1,36 @@
+import os
+
 from telegram import (
     InputMediaDocument,
     InputMediaPhoto,
     InputMediaVideo,
+    LinkPreviewOptions,
     ReplyParameters,
 )
+from telegram.constants import ParseMode
 from telegram.ext import ContextTypes
 from telegram.ext.filters import MessageFilter
+
+# get args from environment virables
+TG_TOKEN = os.environ.get("TG_TOKEN")
+TG_REVIEWER_GROUP = os.environ.get("TG_REVIEWER_GROUP")
+TG_PUBLISH_CHANNEL = os.environ.get("TG_PUBLISH_CHANNEL")
+TG_REJECTED_CHANNEL = os.environ.get("TG_REJECTED_CHANNEL")
+TG_BOT_USERNAME = os.environ.get("TG_BOT_USERNAME")
+TG_RETRACT_NOTIFY = os.getenv("TG_RETRACT_NOTIFY", "True") == "True"
+TG_BANNED_NOTIFY = os.getenv("TG_BANNED_NOTIFY", "True") == "True"
+TG_REJECT_REASON_USER_LIMIT = os.getenv("TG_REJECT_REASON_USER_LIMIT", "True") == "True"
+try:
+    APPROVE_NUMBER_REQUIRED = int(os.getenv("TG_APPROVE_NUMBER_REQUIRED", 2))
+except (TypeError, ValueError):
+    APPROVE_NUMBER_REQUIRED = 2
+try:
+    REJECT_NUMBER_REQUIRED = int(os.getenv("TG_REJECT_NUMBER_REQUIRED", 2))
+except (TypeError, ValueError):
+    REJECT_NUMBER_REQUIRED = 2
+REJECTION_REASON = os.environ.get("TG_REJECTION_REASON", "").split(":")
+TG_DB_URL = os.environ.get("TG_DB_URL", "")
+TG_SINGLE_MODE = os.getenv("TG_SINGLE_MODE", "True") == "True"
 
 
 class PrefixFilter(MessageFilter):
@@ -31,6 +56,8 @@ async def send_group(
 
     # use item_list and type_list to build a list of telegram.InputMediaDocument, telegram.InputMediaPhoto, telegram.InputMediaVideo
     media = []
+    stickers = []
+    gifs = []
     for i in range(len(item_list)):
         match type_list[i]:
             case "photo":
@@ -43,13 +70,32 @@ async def send_group(
                 )
             case "document":
                 media.append(InputMediaDocument(item_list[i]))
+            case "sticker":
+                stickers.append(item_list[i])
+                text = text.strip()
+            case "animation":
+                gifs.append(item_list[i])
+
+    # gifs can not be sent as a group
+    for gif in gifs:
+        sent_messages.append(
+            await context.bot.send_animation(
+                chat_id=chat_id,
+                animation=gif,
+                caption=text,
+                parse_mode=ParseMode.MARKDOWN_V2,
+            )
+        )
 
     for i in range(0, len(media), 10):
         portion = media[i : i + 10]
         if len(portion) > 1:
             sent_messages.extend(
                 await context.bot.send_media_group(
-                    chat_id=chat_id, media=portion, caption=text
+                    chat_id=chat_id,
+                    media=portion,
+                    caption=text,
+                    parse_mode=ParseMode.MARKDOWN_V2,
                 )
             )
             continue
@@ -61,6 +107,7 @@ async def send_group(
                         photo=portion[0].media,
                         caption=text,
                         has_spoiler=has_spoiler,
+                        parse_mode=ParseMode.MARKDOWN_V2,
                     )
                 )
             case InputMediaVideo():
@@ -70,6 +117,7 @@ async def send_group(
                         video=portion[0].media,
                         caption=text,
                         has_spoiler=has_spoiler,
+                        parse_mode=ParseMode.MARKDOWN_V2,
                     )
                 )
             case InputMediaDocument():
@@ -78,8 +126,27 @@ async def send_group(
                         chat_id=chat_id,
                         document=portion[0].media,
                         caption=text,
+                        parse_mode=ParseMode.MARKDOWN_V2,
                     )
                 )
+
+    for sticker in stickers:
+        sent_messages.append(
+            await context.bot.send_sticker(
+                chat_id=chat_id,
+                sticker=sticker,
+            )
+        )
+
+    # if there're only stickers and text message, send text message additionally
+    if not media and not gifs and text:
+        sent_messages.append(
+            await context.bot.send_message(
+                chat_id=chat_id,
+                text=text,
+                parse_mode=ParseMode.MARKDOWN_V2,
+            )
+        )
 
     return sent_messages
 
@@ -99,7 +166,12 @@ async def send_submission(
     # no media or documents, just send text
     if not media_id_list and not documents_id_list:
         sent_messages.append(
-            await context.bot.send_message(chat_id=chat_id, text=text)
+            await context.bot.send_message(
+                chat_id=chat_id,
+                text=text,
+                parse_mode=ParseMode.MARKDOWN_V2,
+                link_preview_options=LinkPreviewOptions(is_disabled=True),
+            )
         )
         return sent_messages
 
@@ -144,6 +216,16 @@ async def send_result_to_submitter(
                 submit_message_id, allow_sending_without_reply=True
             ),
             reply_markup=inline_keyboard_markup,
+            parse_mode=ParseMode.MARKDOWN_V2,
         )
     except:
         pass
+
+
+async def get_name_from_uid(context, user_id):
+    try:
+        user = await context.bot.get_chat(user_id)
+        return user.username, user.full_name
+    except Exception as e:
+        print(e)
+        return "", ""
